user cart

+++ app/Http/Controllers/ProductController.php
@@ -51,8 +51,10 @@ class ProductController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Product $product)
+    public function remove(Product $product)
     {
+        $this->authorize('delete', $product);
+
         $product->delete();
         return response()->json([
             'success' => true
